final 5

+++ src/main/java/com/example/btl_tmdt/controller/admin/AdminCategoryController.java
@@ -1,9 +1,12 @@
 package com.example.btl_tmdt.controller.admin;
 
 import com.example.btl_tmdt.dao.CategoryDao;
+import com.example.btl_tmdt.dao.UserDao;
 import com.example.btl_tmdt.model.Category;
+import com.example.btl_tmdt.model.User;
 import com.example.btl_tmdt.service.CategoryService;
 import com.example.btl_tmdt.service.UserService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,8 +22,18 @@ public class AdminCategoryController {
     @Autowired
     CategoryService categoryService;
 
+    @Autowired
+    HttpSession session;
+    public boolean checkUser(){
+        User user = (User) userService.getUserByUserName((String) session.getAttribute("userName"));
+        return user.getUserRole().equals("2");
+    }
     @GetMapping("")
     public String getCategoryList(Model model){
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         List<Category> categoryList = categoryService.getCategories();
         model.addAttribute("categories", categoryList);
 
@@ -29,6 +42,10 @@ public class AdminCategoryController {
 
     @GetMapping("/add-category")
     public String addCategoryGet(Model model){
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         CategoryDao categoryDao = new CategoryDao();
         model.addAttribute("categoryDao", categoryDao);
         return "admin/category/add-category";
@@ -36,6 +53,7 @@ public class AdminCategoryController {
 
     @PostMapping("/add-category")
     public String addCategoryPost(@ModelAttribute CategoryDao categoryDao, Model model){
+
         Category category = categoryService.getCategoriesByname(categoryDao.getCategory_name());
         if(category != null){
             System.out.println("Category is already existed!");
@@ -48,6 +66,10 @@ public class AdminCategoryController {
 
     @GetMapping("/edit-category/{id}")
     public String editCategoryGet(@PathVariable(name = "id") String id, Model model){
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         CategoryDao categoryDaoToEdit = categoryService.getCategoryById(id).toDao();
         model.addAttribute("categoryDao", categoryDaoToEdit);
         return "admin/category/edit-category";
@@ -56,6 +78,7 @@ public class AdminCategoryController {
 
     @PostMapping("/edit-category/{id}")
     public String editCategoryPost(@PathVariable(name = "id") String id, Model model, @ModelAttribute("categoryDao") CategoryDao categoryDao){
+
 //        Category category = categoryService.getCategoriesByname(categoryDao.getCategory_name());
         Category categoryToEdit = categoryService.getCategoryById(id);
 //        if (categoryToEdit != null) {
@@ -70,6 +93,10 @@ public class AdminCategoryController {
 
     @GetMapping("/delete-category/{id}")
     public String deleteCategory(Model model, @PathVariable(name = "id") String id) {
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         categoryService.deleteCategory(categoryService.getCategoryById(id));
         return "redirect:/admin/category";
     }
