+++ src/main/java/com/example/btl_tmdt/controller/admin/AdminOrderController.java
@@ -2,12 +2,15 @@ package com.example.btl_tmdt.controller.admin;
 
 import com.example.btl_tmdt.dao.OrderDao;
 import com.example.btl_tmdt.dao.ProductInOrderDao;
+import com.example.btl_tmdt.dao.UserDao;
 import com.example.btl_tmdt.model.Order;
 import com.example.btl_tmdt.model.ProductInOrder;
+import com.example.btl_tmdt.model.User;
 import com.example.btl_tmdt.service.OrderService;
 import com.example.btl_tmdt.service.ProductInOrderService;
 import com.example.btl_tmdt.service.ProductService;
 import com.example.btl_tmdt.service.UserService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -31,8 +34,20 @@ public class AdminOrderController {
     @Autowired
     OrderService orderService;
 
+    @Autowired
+    HttpSession session;
+
+    public boolean checkUser(){
+        User user = (User) userService.getUserByUserName((String) session.getAttribute("userName"));
+        return user.getUserRole().equals("2");
+    }
+
     @GetMapping("")
     public String getOrders(Model model){
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         List<OrderDao> orderDaoList = orderService.getOrders();
         model.addAttribute("orderDaos", orderDaoList);
         return "admin/order/orders";
@@ -40,7 +55,10 @@ public class AdminOrderController {
 
     @GetMapping("/order-item/{id}")
     public String getProductInOrder (Model model, @PathVariable(name = "id") String id) {
-
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         Order order = orderService.getOrderById(id);
 
         List<ProductInOrderDao> productInOrderDaos = productInOrderService.getProductInOrder(order).stream()
@@ -58,7 +76,10 @@ public class AdminOrderController {
     @GetMapping("/delete-order/{id}")
     public String deleteOrder (Model model,
                                @PathVariable(name = "id") String id) {
-
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         Order order = orderService.getOrderById(id);
 
         productInOrderService.deleteOrder(order);
