+++ src/main/java/com/example/btl_tmdt/controller/admin/AdminProductController.java
@@ -2,10 +2,14 @@ package com.example.btl_tmdt.controller.admin;
 
 import com.example.btl_tmdt.dao.CategoryDao;
 import com.example.btl_tmdt.dao.ProductDao;
+import com.example.btl_tmdt.dao.UserDao;
 import com.example.btl_tmdt.model.Category;
 import com.example.btl_tmdt.model.Product;
+import com.example.btl_tmdt.model.User;
 import com.example.btl_tmdt.service.CategoryService;
 import com.example.btl_tmdt.service.ProductService;
+import com.example.btl_tmdt.service.UserService;
+import jakarta.servlet.http.HttpSession;
 import jakarta.websocket.server.PathParam;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -29,9 +33,21 @@ public class AdminProductController {
     ProductService productService;
     @Autowired
     CategoryService categoryService;
+    @Autowired
+    UserService userService;
+    @Autowired
+    HttpSession session;
+    public boolean checkUser(){
+        User user = (User) userService.getUserByUserName((String) session.getAttribute("userName"));
+        return user.getUserRole().equals("2");
+    }
 
     @GetMapping("")
     public String getProducts(Model model){
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         List<ProductDao> productDaoList = productService.getProducts().stream().map(Product::toDao).collect(Collectors.toList());
         model.addAttribute("productDaos", productDaoList);
 
@@ -45,6 +61,10 @@ public class AdminProductController {
 //        assert productDaoList != null;
 //        productDaoList.add(productDao);
 //        model.addAttribute("products", productDaoList);
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         ProductDao productDao = new ProductDao();
         List<CategoryDao> categoryDaos = categoryService.getCategories().stream().map(Category::toDao)
                 .collect(Collectors.toList());
@@ -58,6 +78,10 @@ public class AdminProductController {
     public String addNewProductPost(Model model, @ModelAttribute(name = "productDao") ProductDao productDao,
                                     @RequestParam("pictureFile") MultipartFile pictureFile) {
 
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         if (pictureFile.isEmpty()) {
             model.addAttribute("error", "Please upload picture file");
             return "redirect:/admin/product/add-product";
@@ -82,6 +106,10 @@ public class AdminProductController {
 
     @GetMapping("/edit-product/{id}")
     public String editProductGet(@PathVariable String id, Model model){
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         ProductDao productDao = productService.getProductById(id).toDao();
         List<CategoryDao> categoryDaos = categoryService.getCategories().stream().map(Category::toDao).collect(Collectors.toList());
         model.addAttribute("productDao", productDao);
@@ -91,6 +119,10 @@ public class AdminProductController {
 
     @GetMapping("/search")
     public String searchProd(@PathParam(value = "productName") String productname, Model model){
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         List<ProductDao> productDaoListByName = productService.findProductByName(productname).stream().map(Product::toDao).toList();
         model.addAttribute("productDaos", productDaoListByName);
         return "admin/product/products";
@@ -128,6 +160,10 @@ public class AdminProductController {
 
     @GetMapping("/delete-product/{id}")
     public String deleteProduct(@PathVariable(name = "id") String id, Model model){
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         Product productToDelete = productService.getProductById(id);
         productService.deleteProductById(productToDelete);
 //        model.addAttribute("products", productToDelete);
@@ -145,6 +181,10 @@ public class AdminProductController {
 
     @GetMapping("/products-category")
     public String getProductByCategory(@RequestBody CategoryDao categoryDao, Model model){
+        if(!checkUser()){
+            model.addAttribute("userDao", new UserDao());
+            return "client/login";
+        }
         List<ProductDao> productDaoList = productService.getProducts().stream()
                 .map(Product::toDao).toList()
                 .stream().filter(e -> e.getCategoryDao().equals(categoryDao)).toList();
